Removed the user directory path prepending
When you specified a directory for a user's initial directory, it would take the form /home/user/directory with no way to get outside of /home/user for, say, an admin account. Without this prepender, the home directory for your admin account can be /home instead of /home/admin. Downside: you now have to specify an absolute path for your users.